+++ src/lib/st-sidebar/st-sidebar.component.spec.ts
@@ -9,11 +9,12 @@
  * SPDX-License-Identifier: Apache-2.0.
  */
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
-import { ChangeDetectionStrategy } from '@angular/core';
+import { ChangeDetectionStrategy, Component } from '@angular/core';
 
 import { StSidebarComponent } from './st-sidebar.component';
 import { StSidebarItemListComponent } from './st-sidebar-item-list/st-sidebar-item-list.component';
 import { StSearchModule } from '../st-search/st-search.module';
+import { StSidebarItem } from './st-sidebar-item.interface';
 
 describe('StSidebar', () => {
 
@@ -225,4 +226,102 @@ describe('StSidebar', () => {
          }, 0);
       });
    });
+});
+
+
+
+@Component({
+   template: `
+    <st-sidebar [items]="items">
+       <span sidebar-header><span>TEST HEADER</span></span>
+       <span sidebar-footer>TEST FOOTER</span>
+    </st-sidebar>
+      `
+})
+class StSidebarComponentTestComponent {
+   public items: StSidebarItem[] = [
+      { id: 'vault-roles', label: 'Vault Roles' },
+      { id: 'identities', label: 'Identities', class: 'warning', disabled: true },
+      {
+         id: 'masters',
+         label: 'Masters',
+         result: '-',
+         items: [{
+            id: 'sub-item1',
+            label: 'Subitem 1',
+            items: [
+               { id: 'sub-item1.1', label: 'Subitem 1.1', result: '450' },
+               { id: 'sub-item1.2', label: 'Subitem 1.2' }]
+         }, { id: 'sub-item2', label: 'Subitem 2' }]
+      },
+      {
+         id: 'agents', label: 'So long section name', result: '25',
+         items: [
+            { id: 'agents.1', label: 'Subitem 1.1', result: '25' },
+            { id: 'agents.2', label: 'Subitem 1.2' }]
+      },
+      {
+         id: 'agents2', label: 'So long section name 2 without results',
+         items: [
+            { id: 'agents2.1.1', label: 'Subitem 1.1.1', result: '25' },
+            { id: 'agents2.1.2', label: 'Subitem 1.1.2' }]
+      },
+      { id: 'without-children', label: 'No children' },
+      {
+         id: 'sectionA', label: 'Section A',
+         items: [
+            { id: 'sub-sectionA.1', label: 'Subitem 1.1' },
+            { id: 'sub-sectionA.2', label: 'Subitem 1.2' }]
+      },
+      {
+         id: 'sectionB', label: 'Section B', items: [
+         { id: 'sub-sectionB.1', label: 'Subitem 1.1',    disabled: true },
+         { id: 'sub-sectionB.2', label: 'Subitem 1.2' }]
+      },
+      {
+         id: 'sectionC', label: 'Section C',
+         items: [
+            { id: 'sub-sectionC.1', label: 'Subitem 1.1' },
+            { id: 'sub-sectionC.2', label: 'Subitem 1.2' }]
+      },
+      {
+         id: 'sectionD', label: 'Section D',
+         items: [
+            { id: 'sub-sectionD.1', label: 'Subitem 1.1' },
+            { id: 'sub-sectionD.2', label: 'Subitem 1.2' }]
+      }
+   ];
+}
+
+let usageFixture: ComponentFixture<StSidebarComponentTestComponent>;
+let usageComponent: StSidebarComponentTestComponent;
+
+describe('StSidebarComponentComponent in usage', () => {
+   beforeEach(async(() => {
+      TestBed.configureTestingModule({
+         imports: [StSearchModule],
+         declarations: [StSidebarComponent, StSidebarItemListComponent, StSidebarComponentTestComponent]
+      })
+         .overrideComponent(StSidebarComponent, {
+            set: { changeDetection: ChangeDetectionStrategy.Default }
+         })
+         .compileComponents();  // compile template and css
+   }));
+
+   beforeEach(() => {
+      usageFixture = TestBed.createComponent(StSidebarComponentTestComponent);
+      usageComponent = usageFixture.componentInstance;
+   });
+
+   afterEach(() => {
+      usageFixture.destroy();
+   });
+
+   it('It should be able to add some content before and after the menu', () => {
+      expect(usageFixture.nativeElement.querySelector('[sidebar-header]')).not.toBeNull();
+      expect(usageFixture.nativeElement.querySelector('span[sidebar-header]').innerText).toContain('TEST HEADER');
+      expect(usageFixture.nativeElement.querySelector('[sidebar-footer]')).not.toBeNull();
+      expect(usageFixture.nativeElement.querySelector('span[sidebar-footer]').innerText).toContain('TEST FOOTER');
+   });
+
 });