+++ CHANGELOG.md
@@ -8,12 +8,12 @@
 * st-button: Add small button style
 * st-input, st-textarea, st-select, st-radio-button, st-checkbox, st-tag-input: Change disabled style
 * st-modal: Set default max and min width
+* st-sidebar: Allow to add any content from outside to header and footer
 
 **Breaking changes:**
 
 * st-bread-crumb: Changed options from string to object with label and icon to allow icons
 * st-two-list: Includes checkbox all button in list-scroll component. Emits new event with number of elements selected.
-
 
 **Others**